Require that slashing reporters are accounts (#8854)
### Description

Slashing is incentivized with a reward of locked CELO going from the slashee to the reporter. Locked CELO is tied to accounts from the Accounts core contract, so we require that reporters have created an account.

### Other changes

Reworded a description string for an unrelated test.

### Tested

Unit tests.

### Related issues

- Fixes #8828

### Backwards compatibility

This change doesn't change storage layouts or function signatures.