Fixes ODIS matchmaking full-node error handling (#8968)
### Description

The ODIS combiner has recently responded to some matchmaking requests with 500 'unknown error' and this appears to be caused by timeouts reaching Forno. This fix does not attempt to correct the Forno issue, but rather the handling of that error. We should not be responding with 500 here if the error is caused by Forno, and it seems the reason we are is that we return the isVerified promise before it settles, so the error is thrown outside of the isVerified try-catch and ends up being caught by the last-resort 500 'unknown error' try-catch that wraps the entire code path. By using the promise.catch() method, we ensure the correct error handling is used. 

### Other changes

None

### Tested

Will be tested in staging using a dummy full node URL

### Related issues

None

### Backwards compatibility

Yes

### Documentation

None