exclude test files from being compiled into the lib folder. (#8990)
### Description

test files were being include in the lib folders and as such included in any package that used them. confirmed this by looking at the node_modules for use-contract kit and found a bunch of test.js files in there. 

Example
![Screen Shot 2021-11-18 at 2 22 55 PM](https://user-images.githubusercontent.com/3814795/142506286-40ad5d69-e314-4263-8258-348a7009ee21.png)


### Tested

After adding the `exludes` runing `yarn build` in the package folder added the js files but not test files to the lib directory


### Related issues

- Fixes #[issue number here]

### Backwards compatibility

yes
### Documentation

n/a