Skeleton of the new Blockscout staging environment
Provides the main Helm and env variable changes needed for the new rc1staging Blockscout deployment.
As currently running, there are more manual changes that are required to reproduce the setup.
The goal not being to integrate it with celotool but to make it possible to manage without it,
these need to be added in follow-up steps.