Odis forno rate limiting fix (#9083)
### Description

Adds hotfix to combiner code to catch uncaught timeout errors caused by forno rate limiting.
Adds what is hopefully a fix for the forno error handling to the common package. This can be published and tested along with the next combiner release. 

My hypothesis for what is breaking the error handling is that previously we were accessing error.message in the catch block which sometimes fails since error is of type any. By switching to the try/catch syntax we are less likely to have subtle errors like this that go unnoticed. I should be able to test this theory by monitoring the logs during times when forno is causing timeouts. 

### Other changes

None

### Tested

Will be tested via monitoring in staging and alfajores before deploying to mainnet

### Related issues

- NA 

### Backwards compatibility

Yes

### Documentation

None