untangle contractKit intra package dependencies. Make it so nothing inside of the contractKit packgaes needs a kit to be passed into its constructor.
instead use Connection or WrapperCache or move functions close to their source.

This should alllow us to be more selective and decide to one ContractWrapper without needing to initialize a kit which itself would require initializing all bunch of other data.

This should lead to the ability to only import parts of contract kit. and lead to small bundle sizes