Remove oracle check (#9367)
### Description

A check in the Reserve's `addToken` function appeared to be a blocker for getting quicker, single-proposal stable token deployments. The check may have been important in the past, back when the Reserve was responsible for minting/burning stable tokens, and the `isToken` mapping set in the `addToken` function was used for permissioning these behaviors. However, looking at both the monorepo contracts and celo-blockchain, the `isToken` mapping and `_tokens` list of tokens are not used anywhere, thus the check is safe to remove. See [here](https://github.com/celo-org/celo-monorepo/issues/9195#issuecomment-1057942027) for more details.

### Tested

Tests still pass. 

### Related issues

- Fixes #9197