Add io-ts type checks to ODIS domain requests and responses (#9348)
### Description

This PR add `io-ts` type definitions for ODIS Domain requests and responses. This PR utilizes these
schemas to check the received request and response values in the ODIS signer and ODIS client
respectively.

Additionally, this PR completes a planned refactor of `KnownDomain` to `Domain`. The previously
included `Domain` type is removed. It was an attempt to express a constraining type of future
implemented `Domain` subtypes, which is not really a concept implemented by TypeScript.

### Other changes

* Decided I was being far to "clever" with all that `OmitIfNever` malarkey and so removed it. As a
  result, options will now be a required field for requests, even if a domain has none. Instead of
  omitting the field, the options struct will simply be empty in that case.

### Tested

* Added some unit tests for the request schemas.

### Backwards compatibility

Breaking changes to code that is not yet in production.