Refactor (celo-blockchain) fullnode readinessProbe (#9358)
### Description

Refactor the kubernetes readinessProbe check for celo-blockchain fullnode to allow requests when the network is not producing blocks with the expected cadence This should solve having forno "down" when the network is stalled or producing blocks very slowly.

The conditions to consider the node as ready are (any of these):

- LastBlock in node was produced in the last 20 seconds
- The node is peered with more than 0.2*max_full_peers (if that number is between 5 and 30 peers), and the node is not syncing.

### Tested

Tested in private rc1 nodes.