(#9506) Add cronjob template to blockscout helm charts for data migrations
### Description

A cronjob to act as a template for data migration jobs that are to be explicitly and manually triggered by a developer, and run within the cluster (i.e. outside the lifetime of an ssh connection, with logging and metrics). The cronjob is deactivated both by having `suspend: true` set and `schedule: "0 0 31 2 *` ( 31st of February - a non existent date). The suspend property prevents it from being scheduled, but due to schema validation a valid entry is needed for the `schedule` property.

This job is ultimately triggered by a script in the blockscout repository which will be added in a parallel PR. 

### Tested

* Deployed to `rc1staging`, created and ran migration instance

### Related issues

- relates to celo-org/data-services#199

### Backwards compatibility

* It's all backwards compatible