fix(scripts:artifacts): uses `deployedBytecode` instead of `bytecode`
[EIP-170](https://eips.ethereum.org/EIPS/eip-170) puts "a hard cap on the size of an object that can be saved to the blockchain".

Since it's the deployedBytecode that will be stored in the smart contract, and not the initialization bytecode, we should be looking at the size of the deployedBytecode.

That might also explain (parts of) why the on-chain calculation is slightly different, since that uses the eth_getCode RPC method, which returns the deployed bytecode.

With this change the contract size using on-chain method and the build artifact method are equal.