Add FeeCurrencyDirectory to Anvil migrations (#10992)
* refactor(migrations/start_anvil): renames variable to fix typo

* docs(protocol/create_and_migrate): adds comment for context

* docs(protocol/migrations): adds TODOs and comments for context

* chore(protocol/migrations): adds TODO comment

* feat(protocol/migrations): adds `FeeCurrencyDirectory` migration

Not checked that the current changes are sufficient and working.

* feat(protocol/migrations): sets CurrencyConfig for stabletoken

* feat(protocol/migrations_sol): sets example CurrencyConfig

1. adds test oracle address
2. adds test intrinsic gas value

* chore(protocol/migrations_sol): now uses `FeeCurrencyDirectory`, not interface

I was calling `initialize()`, but that function is defined in `FeeCurrencyDirectory` not `IFeeCurrencyDirectory`.

Also updates string length, which was previously forgotten.

At this point the compilation succeeds as expected, and the migration transactions are successful.

* chore(protocol/migrations_sol): adds `MockOracle` to `CurrencyConfig`s

This means, all currencies registered in the `FeeCurrencyDirectory` now support the `getExchangeRate` function, which uses the MockOracle contract under the hood.

* chore(protocol/migrations_sol): removes `TODO` comments

* chore(protocol:SortedOracles.sol): Adds `IOracle` to contract definition

The `SortedOracles.sol` contract already implements the `IOracle.sol` interface. This commit only adds it to the contract definition for ease of reference.

* chore(protocol/migrations_sol): Replaces `MockOracle` with `SortedOracles`

The latest `SortedOracles.sol` implementation implements the `IOracle.sol` interface. That means it implements the required `getExchangeRate()` function.

That means the `MockOracle.sol` contract is not required in the migrations.

* chore(protocol/migrations_sol): removes unused and duplicate imports

* style(protocol/migrations_sol): linting

* test(protocol/test-sol/Validators): updates L2 error messages

Now matches error message defined in `isL2Check.sol`
Source: https://github.com/celo-org/celo-monorepo/blob/668a4050f7ee067948498c333da3feec85fa03fe/packages/protocol/contracts-0.8/common/IsL2Check.sol#L32-L33