chore(workflows): adds `release/core-contracts/` branch condition
The "protocol-test-matrix" is a required check for PRs to `release/core-contracts/**` branches, but the condition that triggers the job could never evaluate be true.

That meant that this part of the workflow could never execute, and would remain "expected" forever, since it's required.

This change ensures the condition evaluates to true, and allows the workflow to run.